pyproject.toml
[tool.poetry]
name = "backend"
version = "0.1.0"
description = ""
authors = ["siknue <siknue0058@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
fastapi = {extras = ["standard"], version = "^0.115.12"}
pydantic = "^2.11.3"
uvicorn = "^0.34.1"
mangum = "^0.19.0"
fastapi-cognito = "^2.9.0"
python-dotenv = "^1.1.0"

[tool.poetry.dev-dependencies]
pytest = "^8.3.5"
pytest-watcher = "^0.4.3"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
